@@ -1,7 +1,7 @@
 {
   "name": "html-minifier-terser",
   "description": "Highly configurable, well-tested, JavaScript-based HTML minifier.",
-  "version": "6.1.0",
+  "version": "7.2.0",
   "license": "MIT",
   "repository": "https://github.com/terser/html-minifier-terser.git",
   "bugs": "https://github.com/terser/html-minifier-terser/issues",
@@ -41,8 +41,9 @@
     "uglify"
   ],
   "engines": {
-    "node": ">=12"
+    "node": "^14.13.1 || >=16.0.0"
   },
+  "type": "module",
   "main": "./legacy/htmlminifier.js",
   "module": "./src/index.js",
   "exports": {
@@ -64,53 +65,50 @@
   ],
   "scripts": {
     "build": "rollup -c",
-    "test:node": "jest",
-    "test:web": "jest --environment=jsdom",
-    "test:watch": "jest --watch",
+    "test:node": "NODE_OPTIONS=--experimental-vm-modules jest --verbose",
+    "test:web": "NODE_OPTIONS=--experimental-vm-modules jest --verbose --environment=jsdom",
+    "test:watch": "NODE_OPTIONS=--experimental-vm-modules jest verbose --watch",
     "test": "npm run test:node",
     "serve": "vite",
     "build:docs": "vite build --base /html-minifier-terser/",
-    "lint": "eslint .",
+    "lint": "eslint . --ignore-path .gitignore",
     "prepare": "is-ci || husky install"
   },
   "dependencies": {
     "camel-case": "^4.1.2",
-    "clean-css": "^5.2.2",
-    "commander": "^8.3.0",
     "he": "^1.2.0",
     "htmlparser2": "^7.2.0",
+    "clean-css": "~5.3.2",
+    "commander": "^10.0.0",
+    "entities": "^4.4.0",
     "param-case": "^3.0.4",
     "relateurl": "^0.2.7",
-    "terser": "^5.10.0"
+    "terser": "^5.15.1"
   },
   "devDependencies": {
     "@babel/core": "^7.16.0",
     "@babel/preset-env": "^7.16.4",
-    "@jest/globals": "^27.3.1",
-    "@rollup/plugin-commonjs": "^21.0.1",
-    "@rollup/plugin-json": "^4.1.0",
-    "@rollup/plugin-node-resolve": "^13.0.6",
-    "alpinejs": "^3.5.2",
     "babel-jest": "^27.3.1",
-    "eslint": "^7.32.0",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.1",
-    "husky": "^7.0.4",
+    "eslint-plugin-promise": "^6.1.1",
+    "@commitlint/cli": "^17.6.6",
+    "@jest/globals": "^29.5.0",
+    "@rollup/plugin-commonjs": "^25.0.2",
+    "@rollup/plugin-json": "^6.0.0",
+    "@rollup/plugin-node-resolve": "^15.1.0",
+    "@rollup/plugin-terser": "^0.4.3",
+    "alpinejs": "^3.12.2",
+    "commitlint-config-non-conventional": "^1.0.1",
+    "eslint": "^8.45.0",
+    "eslint-config-standard": "^17.1.0",
+    "husky": "^8.0.3",
     "is-ci": "^3.0.1",
-    "jest": "^27.3.1",
-    "lint-staged": "^12.1.2",
-    "rollup": "^2.60.1",
-    "rollup-plugin-polyfill-node": "^0.7.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "vite": "^2.6.14"
-  },
-  "benchmarkDependencies": {
-    "chalk": "^4.1.2",
-    "cli-table3": "^0.6.0",
-    "lzma": "^2.3.2",
-    "minimize": "^2.2.0",
-    "progress": "^2.0.3"
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
+    "lint-staged": "^13.2.3",
+    "rollup": "^3.26.0",
+    "rollup-plugin-polyfill-node": "^0.12.0",
+    "vite": "^4.3.9"
   }
 }
